+++ src/index.scss
@@ -1,0 +1,3 @@
+body {
+    font-family: 'Manrope', sans-serif;
+}
