--- src/pages/homepages/homepage.components.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import Directory from '../../components/directory/directory.component';
-import "./homepage.styles.scss"
-// import Example from "../react/react"
-
-const HomePage = () => (
-  <div className='homepage'>
-    <Directory />
-    {/* <Example /> */}
-  </div>
-);
-
-export default HomePage;
