--- src/pages/homepages/homepage.styles.scss
@@ -1,6 +1,0 @@
-.homepage {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px 80px;
-}
